Fixes to get Gifts fully working
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@5184 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/busrules/CollectingEventBusRules.java
@@ -82,7 +82,7 @@ public class CollectingEventBusRules extends AttachmentOwnerBaseBusRules
                 
             } else
             {
-                isOK = okToDelete(0, new String[] {"collectionobject", "CollectingEventID"}, (Integer)ce.getId());
+                isOK = okToDelete(0, new String[] {"collectionobject", "CollectingEventID"}, ce.getId());
             }
             deletable.doDeleteDataObj(dataObj, session, isOK);
